+++ IOCv2.IntegrationTests/Factories/TestDbSeeder.cs
@@ -30,7 +30,7 @@ public static class TestDbSeeder
             TaxCode = "0101248141",
             Industry = "IT Test",
             Address = "Test Address",
-            Status = (short)EnterpriseStatus.Active
+            Status = EnterpriseStatus.Active
         });
 
         context.SaveChanges();
